+++ Scripts/Python/payiUrwinBrain.py
@@ -430,7 +430,7 @@ class payiUrwinBrain(ptResponder):
         PtDebugPrint("payiUrwinBrain: Generated a valid spawn time: %d" % (firstTime))
         spawnTimes = [firstTime]
 
-        while isinstance(spawnTimes[-1], long):
+        while isinstance(spawnTimes[-1], int):
             randnum = random.randint(kMinimumTimeBetweenSpawns, kMaximumTimeBetweenSpawns)
             newTime = spawnTimes[-1] + randnum
             if newTime < endOfToday:
